Tiny fixes

+++ dotfiles/virtualbox/restore.sh
@@ -2,6 +2,9 @@
 
 source ../_common.sh
 
-wget https://download.virtualbox.org/virtualbox/rpm/fedora/virtualbox.repo -O /etc/yum.repos.d/virtualbox.repo
+if [ ! -e /etc/yum.repos.d/virtualbox.repo ]
+then
+	wget https://download.virtualbox.org/virtualbox/rpm/fedora/virtualbox.repo -O /etc/yum.repos.d/virtualbox.repo
+fi
 
 dnf_install VirtualBox-6.1
